add VALUE_TYPE_MINMAX type and xml generator logic
If an item of this value type is linked to another position:ID, use the
min/max/type/labels from the linked item when generating xml.

Eg, if this is 8772:28 (LINK_MIN:LINK_POSITION_1), then
look up value from 8770:28 (LINK_TO:LINK_POSITION_1).
If the result is 2626:18 (PRESET_LEVEL), then use the 'values'
data from there to generate a label for the original param->value.